refactor: Number -> Denomination 도메인명 수정
