+++ src/main/java/blackjack/domain/card/Deck.java
@@ -13,7 +13,7 @@ public class Deck {
     static {
         final Stack<Card> pack = new Stack<>();
         final List<Card> cards = Arrays.stream(Suit.values())
-                .flatMap(suit -> Arrays.stream(Number.values()).map(number -> new Card(number, suit)))
+                .flatMap(suit -> Arrays.stream(Denomination.values()).map(number -> new Card(number, suit)))
                 .collect(Collectors.toList());
         pack.addAll(cards);
         TRUMP = pack;
