+++ src/test/java/blackjack/domain/card/DeckTest.java
@@ -1,6 +1,6 @@
 package blackjack.domain.card;
 
-import static blackjack.domain.card.Number.ACE;
+import static blackjack.domain.card.Denomination.ACE;
 import static blackjack.domain.card.Suit.SPADE;
 import static org.assertj.core.api.Assertions.assertThat;
 import static org.assertj.core.api.Assertions.assertThatThrownBy;
